Fix list item icon position
The position of icon indicating whether a list item is added manually
is fixed for empty list items. Further, the list item is extracted into
a separate component.

Fix: #1374
